library/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
orm-choices==1.0.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
